GKI: clk: qcom: npucc-lito: handle probe errors correctly
Currently, each time a probe deferral occurs for npucc-lito,
a new refcount is placed on the vdd_cx level. This was preventing
the device from entering into the cxsd sleep state. This fix
makes sure that when the probe get deferred(or fails otherwise),
any refcounts on voltage levels already placed by the driver
are removed.

Bug: 147116623
Test: build, boot test, all devices still probe, device enters cxsd
mode.
Change-Id: I2b3946fcf5c92414564d32bbed36087f4641e9d2
Signed-off-by: Hridya Valsaraju <hridya@google.com>