scsi: ufs-qcom: use downstream driver for SCM calls
There are two SCM call drivers in this kernel: one from the upstream
kernel in drivers/firmware/qcom_scm.c, and a downstream driver in
drivers/soc/qcom/scm.c.

Unfortunately, trying to make the ICE SCM calls with the upstream driver
crashes with this kernel on bramble and QRD8250, seemingly due to
downstream changes to DMA, and maybe due to other reasons too.

So switch ufs-qcom to use the downstream driver instead, which is what
the old ICE implementation used.

Bug: 147259927
Change-Id: I18ba7161b7cb7b5a8b01bee28078392093487624
Signed-off-by: Eric Biggers <ebiggers@google.com>