GKI: watchdog_v2: ensure watchdog is loaded first
Noticed phy-qcom-ufs loads first, so we added a softdep in that
driver to require watchdog_v2 to load prior to it.

This solved the problem today, and may require more love in the
future if another driver takes the mantle of first loaded.  The
final solution, however, is that watchdog_v2 is manually specified
in modules.order as part of build process for the ramdisk boot.img.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 142496416
Test: confirm on boot that watchdog is loaded first
Change-Id: I868c155a685896f373598db890b421230d7cf99f