regulator: tps: Rename ambiguous DT property
Properties ending with -gpio and -gpios are common DT bindings used to
point to gpio phandles. The nr-control-gpios property is just trying to
list a number and not point to a gpio.

of_devlink feature parses common DT bindings to figure out device
dependencies. This ambiguously name property can cause a lot of warnings
in the logs when of_devlink is enabled because the property is not
actually pointing to a GPIO.

To avoid the unnecessary warning, rename the property to a less
ambiguous name.

Signed-off-by: Saravana Kannan <saravanak@google.com>
Change-Id: I6202c30c5d770d365f985de3ac5b41ef1550d44e