[TEMPORARY] ufs-qcom-ice: support legacy wrapped keys solution
Temporarily add a kconfig option to make the ICE driver not fix the byte
order of keys passed to ICE, and set it on kona and lito.

This is needed for compatibility with unmodified QCOM userspace builds,
which use QCOM's legacy wrapped keys solution.

This commit should be reverted once all needed fstabs have been updated.

Bug: 147259927
Change-Id: Iedab7d3dac044ecb80cc9cb0b9053549c09395d0
Signed-off-by: Eric Biggers <ebiggers@google.com>