GKI: msm: early_random: enable QCOM_EARLY_RANDOM tristate
This module needs to load early if it is not built in.  As a module
it will add entropy upon being loaded.  If built in will be called
directly at the appropriate point in the kernel startup.

The softdep order has to be maintained in all modules requiring it,
and must be synchronized.  The modules that needed to be touched in
all future adjustments so far are:
- drivers/phy/qualcomm/phy-qcom-ufs.c (already)
- drivers/irqchip/qcom-pdc.c
- kernel/trace/msm_rtb.c (especially if enabled for triage/debugging)

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 148665638
Bug: 142939781
Test: module instantiates and contributes to entropy
Change-Id: I54b69572cc0abe585d04c5d821065db106998d4f