GKI: soc: qcom: pil: Do additional error checking
Since the qcom,msm-bus,name property is being used to indicate
dependency on the ad-hoc-bus device without actually setting the rest of
the bus related properties (because they won't be used), we need to
improve the error checking. Otherwise, the driver tries to read rest of
the bus properties and fails.

Bug: 143786221
Tested: Booted and made sure same devices probe
Tested: Unlock the phone and play with home screen
Change-Id: I625f0785d2542b2c5d19427f5d000aff3736bd5d
Signed-off-by: Saravana Kannan <saravanak@google.com>