GKI: soc: qcom: ssr: Rename module to subsystem-restart
The module params are set from userspace and userspace expects the
module parameters to be available under:
/sys/module/subsystem_restart/parameters/

The C file is called subsystem_restart.c. But we need to combine
sysmon-qmi.o into subsystem_restart.ko. But that's not possible as far
as I can tell without renaming subsystem_restart.c or
subsystem_restart.ko. So, rename the module to subsystem-restart.ko.

The tricky part here is that all module names have the "-" changed to
"_". So, this keep the module name same for userspace while allowing to
build the module without having to rename the C file.

Bug: 149531990
Signed-off-by: Saravana Kannan <saravanak@google.com>
Change-Id: If1eb1d9d2cfede7aa48a41fb9e2a4ec9e6a3bd4c