GKI: msm: msm_bus: fix trace events to avoid duplicates
When the msm-bus driver was modularized, the trace points were exported
in both the msm_bus.o and msm_bus_dbg_rpmh.o. This caused overlap
between the traces since it exported the same trace events in both
modules which lead to the below crash. This patch splits up the trace
events between the two modules.

[   71.072507] c7      1 Unable to handle kernel paging request at virtual address ffffffa3dd1537a4
...
[   71.601712] c7      1 Call trace:
[   71.605116] c7      1  0xffffffa3dd1537a4
[   71.609237] c7      1  event_trace_add_tracer+0x6c/0x12c
[   71.614696] c7      1  instance_mkdir+0x138/0x288
[   71.619524] c7      1  tracefs_syscall_mkdir+0x48/0x78
[   71.624793] c7      1  vfs_mkdir2+0xf0/0x174
[   71.629182] c7      1  do_mkdirat+0x94/0x12c
[   71.633561] c7      1  __arm64_sys_mkdirat+0x24/0x30

Signed-off-by: Will McVicker <willmcvicker@google.com>
Bug: 150484986
Test: compile, boot bramble with both GKI and non-GKI kernel
Change-Id: I28a59d949d5e27e0e8cefaa1b9d19f2d1cbf8229