msm_geni_serial: Correct the DMA RX interrupt logic
Correct the DMA RX interrupt handling logic to check for
RX_DMA_DONE done as sometimes RX_EOT bit may not be set for
a successful transfer. But, the RX_DMA_DONE will always be
set for a transfer. This issue is caused by the change
with commit : 6a0ff35cad
Disable and enable interrupt in pm runtime suspend,
pm runtime resume respectively for uart so that there
will not be any spurious core interrupt occurring when the
driver is in suspend state.

Bug: 176298527
Bug: 192129525
Test: manual

Change-Id: I9074d9f373d08982cb82af50a36289061962cf2d
Signed-off-by: Prudhvi Yarlagadda <pyarlaga@codeaurora.org>