ANDROID: GKI: Make linux_banner a C pointer
This variable can have its ABI intentionally changed by anybody building
on a different hostname etc. so it does not make sense to monitor it as
a C string. Make linux_banner a C pointer so it is checked for presence
but not for length.

Bug: 154345163
Test: boot bramble, verify list of probed devices
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I932e7c2ca1a1bd6d730462249005515c71771ed8
(cherry picked from commit 44d67b57923f6bf4b98f49f5aa66cdbeebb449b4)
Signed-off-by: Will McVicker <willmcvicker@google.com>