GKI: UPSTREAM: gcc-10 warnings: fix low-hanging fruit
commit 9d82973e032e246ff5663c9805fbb5407ae932e3 upstream.

Due to a bug-report that was compiler-dependent, I updated one of my
machines to gcc-10.  That shows a lot of new warnings.  Happily they
seem to be mostly the valid kind, but it's going to cause a round of
churn for getting rid of them..

This is the really low-hanging fruit of removing a couple of zero-sized
arrays in some core code.  We have had a round of these patches before,
and we'll have many more coming, and there is nothing special about
these except that they were particularly trivial, and triggered more
warnings than most.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Change-Id: I728287bcc9974ef5a14dab6a8f2688130a74af22
Bug: 157180483
(cherry picked from commit 51cc5495ff35519ba45d02acbf15999be680d428)
Signed-off-by: Will McVicker <willmcvicker@google.com>