fscrypt: remove unnecessary extern keywords
Remove the unnecessary 'extern' keywords from function declarations.
This makes it so that we don't have a mix of both styles, so it won't be
ambiguous what to use in new fscrypt patches.  This also makes the code
shorter and matches the 'checkpatch --strict' expectation.

Link: https://lore.kernel.org/r/20200511191358.53096-4-ebiggers@kernel.org
Signed-off-by: Eric Biggers <ebiggers@google.com>

(cherry picked from f2fs-stable/linux-4.14.y
 commit de635b5d05e152b4d1d64b17b63f28e4e88959cb)

Conflicts:
	fs/crypto/fscrypt_private.h
	include/linux/fscrypt.h

Bug: 154167995
Change-Id: I9224c9efbc4a534d104e78ed6a49399d2c2836ee