fs-verity: remove unnecessary extern keywords
Remove the unnecessary 'extern' keywords from function declarations.
This makes it so that we don't have a mix of both styles, so it won't be
ambiguous what to use in new fs-verity patches.  This also makes the
code shorter and matches the 'checkpatch --strict' expectation.

Link: https://lore.kernel.org/r/20200511192118.71427-3-ebiggers@kernel.org
Signed-off-by: Eric Biggers <ebiggers@google.com>

(cherry picked from f2fs-stable/linux-4.14.y
 commit bc990c874500cb2fc2cbc377f9854562ff309b5d)
Bug: 154167995
Change-Id: I49c4b7af819b785682b696f035b47532c067eb13