google_ttf: fix estimates while AC is on
Use the health vtier when AC is enabled and fix a bug in the
calculation of the ratio when available power is under soc
demand.

Also remove the trigger on rest voltage (not used), simplify
code and interface, log health charging state to the ttf
logbuffer and add sysfs nodes to check trigger and time to ac.

Use simple_strtol() when parsing the deadline

Bug: 163087907
Test: patchwork, device
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Ieee150c62c80ff75f1688a53ab1508ef321e7c44