ANDROID: Update ABI representation for libabigail update
A recent update of libabigail introduces some more tracking options
(ELF variable sizes, aliased symbols, CRC values). Unfortunately this is
incompatible with the current representation. Hence update.

Note, the elf-symbol are moved from the modules to the vmlinux in this
version. This does not harm the analysis. It will be corrected in a
later version again.

Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: If363c5629a94844e35a9c7de87f89f2acd091a81