msm: camera: cci: Fix incorrect use of cci config ioctl
The cci configuration will be transitioned to a new API that does not
require routing through the v4l layer. This is work-in-progrss so in
the mean time prevent the device from being exposed as configurable
from userspace.

The ioctl will still be exposed to kernel users so fix the arg size
as well. We want size of struct not pointer.

CRs-Fixed: 2702760
Change-Id: I9c7bd8f76980603dbf27e1c5bc9b19f8a3b8a39a
Signed-off-by: Fernando Pacheco <fpacheco@codeaurora.org>
Signed-off-by: VijayaKumar T M <vtmuni@codeaurora.org>

Bug: 170138526
Test: Build Pass, Camera CTS test, camera basic function test
Change-Id: I5bb4b863cc4360b37e4819a4afc1502d802111b8
Signed-off-by: Pierce Huang <piercehuang@google.com>