Prevent rollover in ODPM energy values
Bug: 158706436

Use 128-bit arithmetic to compute the ODPM energy values.

Signed-off-by: Lee Shombert <shombert@google.com>

Test: Built a debug image that computed energy using the old logic and
the new logic.  The debug image also emitted the hardware registers so
that the output could be postprocessed on a computer with full 128-bit
support.  Verified that the new code generates the same result as the
old code until rollover.  Verified that the new code generates the same
result as 128-bit arithmetic, for the full 48 bits of the odpm chip
accumulator.

Change-Id: Ibb0dd6a4800daf4ba53723036384f93b6635c100