dm: dm-user: New target that proxies BIOs to userspace
dm-user is essentially FUSE for block devices: as BIOs come in through
device mapper they are proxied to a userspace daemon via a control misc
device.

This is very much a work in progress.  There's a handful of FIXMEs
spread throughout the code with more details.  As far as I know there is
nothing broken with the current code, there's just more work to do.

Test: Ran the selftests on the version of this I'm developing for Linus'
      tree, on both 4.19 and 5.8.
Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>
Change-Id: If5bcd8a43c3db5b556563ba303f474dd0d2902e8