Revert "ANDROID: Add OWNERS files referring to the respective android-mainline OWNERS"
This reverts commit 9790f49ad80eb7759f7e70b33cd64e2c3fd4a927. Because
the partner gerrit does not have visibility into android-mainline, it
can't find these OWNERS. This shouldn't be a problem, but a bug in
gerrit is causing patches to block on this nonexistent file. Revert this
until a proper fix is in place.

Bug: 197928983
Change-Id: I918ade6b9482d9edff321b01d5d86509e0e87336
Signed-off-by: J. Avila <elavila@google.com>
(cherry picked from commit ea799c3d27e971fa9cdeeaee4858324f070561c6)