pd_engine: Implement limit sink current
This implements the kernel changes of the new API introduced in Android
T that limit the current draw when the device acts as Sink. Userspace
processes like USB HAL execute the APIs by writing to the device
attributes which will then vote the values to ICL votables.

usb_limit_sink_current has to be set before the write
to usb_limit_sink_enable. usb_limit_sink_current will not be
updated if usb_limit_sink_enable is already enabled.

Bug: 201006190
Signed-off-by: Kyle Tso <kyletso@google.com>
Change-Id: I91634dbec0fa7824757c0f08ddc61abf1f721d06