serial: msm_geni_serial: Fix console log false start issue
With the commit 7dca9fa41c18 after merging android-4.9-q
(4.9.292), it is possible to re-configure the preferred console
when the serial ports are added even if the cmdline parameter
'console' is set as null.

Set the flag 'console_set_on_cmdline' as 1 directly since
we do not need to setup the console at runtime.

Bug: 211946493
Bug: 210364486
Change-Id: I2f472520ad5ff8593878cfd4f818febc965d828e
Signed-off-by: JohnnLee <johnnlee@google.com>