treewide: Add core-kernel owners and define core kernel files
Add members of the Android Common Kernel (ACK) team so that
any changes to core kernel files can be prevented. The pixelwatch
bsp team will maintain ownership of all Qualcomm files for Monaco
in this project.

The OWNERS files in this commit were auto-generated using the
generate_owners.sh script after compiling a debug build, since
that has all of the files that Monaco would use.

Bug: 261218518
Change-Id: Id2f87892602886060c6ae60c929e25d18579f0ad
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>