include: Fix OWNERS file
The OWNERS file for the include directory incorrectly configures
ownership of everything under include/ to the accounts listed under
OWNERS_pixelwatchbsp, which was not intended.

The original intent was to give ownership of only the files listed
in include/OWNERS to the accounts listed under OWNERS_pixelwatchbsp,
so fix the OWNERS file to reflect this.

Bug: 261218518
Change-Id: I259f32a4a2bb007d80c13b3c7cf815072ffc6fb9
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>