include: Fix pattern globs for applying ownership rules
Some of the per-file restrictions are being applied to all
files under include/. This is happening because each per-file
restriction is applied to the list of comma separated globs
specified in the same line. For example:

per-file dt-bindings/clock/qcom,*=set noparent
per-file dt-bindings/clock/qcom,*=file:/OWNERS_pixelwatchbsp

gives ownership of dt-bindings/clock/qcom and all files under
include/ (because all file paths match the * regex glob) to
OWNERS_pixelwatchbsp, which is not what was intended.

Fix the per-file restrictions that are specifying * as a glob by
removing the comma in those lines to ensure that the correct glob
that was intended is used.

Bug: 261218518
Change-Id: Ieba37ba670058e842466c715596a945827446af0
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>