dcvs: increase size of mbps_zones
Increase the number of zones from 10 to 11 to handle the number of
zones necessary.

to_mbps_zone checks both the list length, and for mbps_zones[i], so
modify store_list_attr to only set node->name[i] = 0 when the number of
entries in the list is smaller than the max list size.

Bug: 271919927
Change-Id: I55fa1d16b1366748233ba300496d92e59f09d694
Signed-off-by: Ben Fennema <fennema@google.com>