power supply: Fix erroneous positive return from pmic_voter_compat_cb
The calling interface is using simple attribute which expects a return 0
code. If we return a positive non-zero code we observe that attempts to
update the value via the sh echo command such as:

echo 1 > /d/gvotables/USB_ICL/force_int_active

result in a FORTIFY error in the function c_print() within the shell sh.
The c_print() function makes a call to write() and assumes the return
value will not be greater than the nbyte argument used in the call to
write(). This will not hold true and result in integer underflow.

Bug: 262324032
Test: echo 1 > /d/gvotables/USB_ICL/force_int_active returns 0
Change-Id: I7f3e595a65f66220acde60b817a8cd372d52e64c
Signed-off-by: Brandon Edens <brandonedens@google.com>