arm64: defconfig: Sync gki_defconfig to ACK's gki_defconfig
The ARM64 gki_defconfig file in the device kernel is out of sync
from the one used to compile the GKI kernel. This is not ideal,
as the device kernel's copy of the gki_defconfig is combined with the
device specific config fragment when building the device kernel modules
to ensure that the modules are compatible with the GKI kernel. So, align
the device kernel's gki_defconfig file to the ACK gki_defconfig file.

None of the device kernel drivers use the extraneous config that
this patch removes from the gki_defconfig, so this patch has no
functional change.

Bug: 280100152
Change-Id: I1165f0cebd5935b945b570ac984c3baf72667cce
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>