qcom_glink_native: Eliminate "misconfigured IRQ" from wakeup reason
irq_may_run() in kernel/irq/chip.c considers the combination of
IRQF_NO_SUSPEND and enable_irq_wake() to be invalid. That is not my
understanding of
https://www.kernel.org/doc/Documentation/power/suspend-and-interrupts.rst
but "misconfigured IRQ" is unsettling, so do as the code commands. Worst
case, the kernel does a bit more work to complete suspend, only to
immediately wake again.

Bug: 256019411
Change-Id: I17b34a8e102784318adbe37b46108eef644ef69c
Signed-off-by: Andrew Evans <andrewevans@google.com>