qcom-mpm: Stop mpm from replaying TZ interrupts as AP wakes (for real)
Commit f9fa0d0aeb405bcbe5aff958180a5848df78568e was accidentally rebased
on itself, rendering it a null change. Here is the actual patch from
QCOM.

andrewevans@ note: From what I've been able to deduce, this patch
disables (or rather doesn't enable) the MPM GIC chip as a wake
source. Radu has demonstrated that most of the wakes on this IRQ are
unactionable (no active or pending interrupt to service), and the few
that aren't are handled by other IRQ handlers that are wake enabled.

Bug: 262736214
Change-Id: I325d6147378b83c232db95f29fbda2c7eb4b8ae8
Signed-off-by: Radu Solea <radusolea@google.com>
Signed-off-by: Andrew Evans <andrewevans@google.com>