qti-qbg: Promote the internal SOC state values to info log.
When the system updates the SOC values we want to know the state of the
precursor SOC values in the system. This patch promotes a debug log to
info adding a single line each time system receives an update from the
PMIC which varies in frequency based upon the battery gauge measurement power state.

Bug: 278938076
Test: Confirm we now have qti-qbg logged SOC values.
Change-Id: I190b005e35a469185c66ac50c6249b71806a0fec
Signed-off-by: Brandon Edens <brandonedens@google.com>