spi: spi-msm-geni: Notify SPI core about suspend and resume
During PM suspend, SPI controller went on suspend and when the SPI
client drivers invoke spi_sync() to complete the transfer, it will block
forever, resulting to crash.

With this implementation SPI controller notifies to SPI core
that it is suspended or resume. So, if any spi_sync() call during
suspend will return error instead of waiting.

Change-Id: Ic4c4ee3a0bab90712453618e76455b668279f691
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>