clk: qcom: smd-rpm: Add support for clock handoff
During boot-up if the clock driver put across a initial high vote then
there is no way to remove these votes which can cause the corresponding
votes to be left always high till the CPU suspends causing power
regression. Thus introduce a device property which can remove the
initial votes from the clock driver and leave it up to the client to
keep a vote.

Change-Id: I068c87d06df908ee0ed0e58f69d46933b301ea6f
Signed-off-by: Taniya Das <quic_tdas@quicinc.com>
Signed-off-by: Prerna Singh <quic_prersing@quicinc.com>