mmc: sdhci-msm: Handle multiple calls to vendor_hook functions
The vendor_hooks are being registered in probe (instead of in init), so
if multiple devices meet the criteria being checked (sdio, nonremovable,
etc.) the functions registered can be called multiple times. Also, the
functions are registered after checking a specific criteria, but with
multiple devices, one device could meet the criteria and another one
could not. This means the original registration criteria needs to be
checked at the beginning of the vendor_hook function and the function
should return immediately if the check fails.

partial_init takes in a boolean to indicate if the init has occurred
already. If it is set, the function should immediately return (as
calling mmc_parital_init multiple times leads to errors).

A further optimization would be to ref count the occurrences of
MMC_CAP_NONREMOVABLE so the calls to register_trace_android only occur
once and the vendor_hook is not guaranteed to be called twice, with the
second immediately returning, when two nonremovable devices exist.

Bug: 302274804
Change-Id: I12b8c597128a167810cd58c305e444ca4b762180
Signed-off-by: Ben Fennema <fennema@google.com>