stm: Add __nocfi to stm functions and use rcu notrace functions
STM drivers are built as modules. When CFI is enabled, __cfi_slowpath
will be called when functions are called among stm drivers. If lock
event is enabled at the same time, as rcu_lock_acquire function will
be called in __cfi_slowpath and stm functions, there will be a circular
callstack. Add __nocfi to stm functions and use rcu notrace functions
to fix this issue.

Change-Id: Ie62c98775e45099ae635c9d94c0d3615cc17e942
Signed-off-by: Xiaoqi Zhuang <quic_xiaozhua@quicinc.com>