smblite: Add notifier for usbin plugin
Other modules can subscribe to this notification to know when the
USB was connected / disconnected. This is the most precise way to
subsribe to these events. However, subscribers need to take care not
to do too much work in their notification functions as the notifier
happens inside the usbin-plugin threaded IRQ handler.

Bug: 308351023
Change-Id: I81d2b8d84d8d30d5b322075a5b18e35a127d516b
Signed-off-by: Andrei Ciubotariu <aciubotariu@google.com>