smblite: Support runtime APSD toggling
Using a new "apsd_en" sysfs node, we can turn APSD on/off.
This can be useful when decoupling the USB PHY data functionality from
VBUS. smblite uses the regulator device interface to turn on the
data line "regulator", which is handled by phy-msm-snps-hs and in
reality turns on the PHY and puts the data lines in a special
non-driving (Hi-Z) mode, for APSD. However, if the PHY is currently
spoofed on, this mode will render the PHY inoperative, since the
usb shim spoof will not be toggled off then on again to put the PHY
back into functional mode.

Bug: 320560592
Change-Id: I63ace561e5c6e6924726eca01fa03c7440a1d261
Signed-off-by: Andrei Ciubotariu <aciubotariu@google.com>