arm64: defconfig: Add "# nocheck:" lines to unnecessary config options
Certain config options from monaco_GKI.config are disabled when the
sw5100.fragment is overlayed on top. This causes the kernel build tools
to complain because it will compare each config fragment against the
final defconfig file that is used to compile the device kernel modules
to ensure tha the values are what is expected. However, this doesn't
seem to be working properly, so add "# nocheck:" lines to bypass
the checks for now.

Bug: 322530495
Bug: 322408856
Change-Id: Ifcc6ce42343e557a1fc7096a8727279c7254f3e6
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>