[DO NOT MERGE][PATCH] serial: msm_geni_serial: Update secondary command register settings
In RX sequencer when the RX line is low, at start
of RX if 16 cycles of low is detected it is
considered as a start bit and if at the 10th bit
high is not detected it becomes a break start and
also a framing error.

Due to this 0x00 is present in the starting of RX
fifo. This will not trigger a stale interrupt as
stale counter starts on a proper frame completion
and for a complete frame.

To avoid 0x00 at start of RX fifo, configure secondary
command register param fields for character with framing
error & break character is not written in RX fifo.

Bug: 304874794
Test: BCST
Signed-off-by: Visweswara Tanuku <quic_vtanuku@quicinc.com>
Signed-off-by: Prasanna S <quic_prass@quicinc.com>
Signed-off-by: Vance Yu <vanceyu@google.com>
Change-Id: Ib7826d09cb0467467fdc6b0f46d80ef1b264ee7a