qti-qbg: Toggle PMIC-side interrupt-enable for FIFO full, during init
If the QBG SDAM FIFO fills up before IRQ registration, the interrupt
that the PMIC raises won't be acted upon, and so the FIFO never gets
cleared, and the PMIC won't raise the interrupt again.

To work around this, disable the interrupt from the PMIC side before
registering the IRQ handler, then re-enable the interrupt from the
PMIC side aftewards. This is done by writing 0x00/0x80 to the
QBG_SDAM_INT_TEST1 register.

This toggling ensures the PMIC raises the interrupt if FIFO is
currently full, no matter when it ended up filling up. Furthermore,
this approach keeps the IRQ handler as the only piece of code that
clears the FIFO after IRQ registration, avoiding concurrency concerns.

Bug: 331463612
Change-Id: I45f0f6b7ad7a350246f44c1a30ef29cda5dbd420
Signed-off-by: Andrei Ciubotariu <aciubotariu@google.com>