qcom-spmi-adc5-gen3: BUG_ON() on extended ADC failures
On multiple occations a device becomes unable to access the ADC on the
PMIC for multiple hours. This situation only recovers once the device
reboots.

We currently have no steps to repro, so crashing the system at that
point will give more information about the issue.

This is also a stopgap measure for user safety. Some of the thermal
sensors are connected to this ADC. When the system gets in this failure
state, it is unable to monitor those sensors for a long time. Forcing a
reboot will allow thermal protection to resume.

Bug: 329331988
Change-Id: I35e4b76bcc817b8a0432598b7fdfacf9f4742fa2
Signed-off-by: Daniel Price <danielprice@google.com>