qpnp-smblite: Log interesting PM5100 registers early during probe
Add an array of registers to be logged very early on in probe, before
they are likely going to be overriden by the rest of initialization.

The logs can be used to determine if the bootloader/OTP PMIC settings
are fine during the window where the linux kernel is booting but
the smblite module has not probed.

Bug: 342905505
Change-Id: Ib1aec3cb06dcd4a29714daa49237687e45fec484
Signed-off-by: Andrei Ciubotariu <aciubotariu@google.com>